Add HTTP executor to flow
The HTTP executor runs tasks in process, like the Fn executor, however
it starts tasks by HTTP POSTing a JSON representation of the task to
a specified task server URL.

It is expected that an entrypoint using the HTTP executor will start an
HTTP server that parses incoming tasks and invokes the handler. In
future the executor itself should probably control the server, and
runTask should take an HTTP request as input.